+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 1c3027293ac9cb8a0d03b18c9ca447d62c2733d7
+// Checksum: 358960b72c6e4f48d3e2e9d52be3abbe9e8b5a9f
 //
 // ignore_for_file: unused_import
 
@@ -426,6 +426,19 @@ final class _EvaluateVisitor
         });
       }, url: "sass:meta"),
 
+      BuiltInCallable.function("module-mixins", r"$module", (arguments) {
+        var namespace = arguments[0].assertString("module");
+        var module = _environment.modules[namespace.text];
+        if (module == null) {
+          throw 'There is no module with namespace "${namespace.text}".';
+        }
+
+        return SassMap({
+          for (var (name, value) in module.mixins.pairs)
+            SassString(name): SassMixin(value)
+        });
+      }, url: "sass:meta"),
+
       BuiltInCallable.function(
           "get-function", r"$name, $css: false, $module: null", (arguments) {
         var name = arguments[0].assertString("name");
@@ -450,6 +463,20 @@ final class _EvaluateVisitor
         if (callable == null) throw "Function not found: $name";
 
         return SassFunction(callable);
+      }, url: "sass:meta"),
+
+      BuiltInCallable.function("get-mixin", r"$name, $module: null",
+          (arguments) {
+        var name = arguments[0].assertString("name");
+        var module = arguments[1].realNull?.assertString("module");
+
+        var callable = _addExceptionSpan(
+            _callableNode!,
+            () => _environment.getMixin(name.text.replaceAll("_", "-"),
+                namespace: module?.text));
+        if (callable == null) throw "Mixin not found: $name";
+
+        return SassMixin(callable);
       }, url: "sass:meta"),
 
       BuiltInCallable.function("call", r"$function, $args...", (arguments) {
@@ -522,7 +549,32 @@ final class _EvaluateVisitor
             configuration: configuration,
             namesInErrors: true);
         _assertConfigurationIsEmpty(configuration, nameInError: true);
-      }, url: "sass:meta")
+      }, url: "sass:meta"),
+      BuiltInCallable.mixin("apply", r"$mixin, $args...", (arguments) {
+        var mixin = arguments[0];
+        var args = arguments[1] as SassArgumentList;
+
+        var callableNode = _callableNode!;
+        var invocation = ArgumentInvocation(
+          const [],
+          const {},
+          callableNode.span,
+          rest: ValueExpression(args, callableNode.span),
+        );
+
+        var callable = mixin.assertMixin("mixin").callable;
+        var content = _environment.content;
+
+        // ignore: unnecessary_type_check
+        if (callable is Callable) {
+          _applyMixin(
+              callable, content, invocation, callableNode, callableNode);
+        } else {
+          throw SassScriptException(
+              "The mixin ${callable.name} is asynchronous.\n"
+              "This is probably caused by a bug in a Sass plugin.");
+        }
+      }, url: "sass:meta", acceptsContent: true),
     ];
 
     var metaModule = BuiltInModule("meta",
@@ -1730,40 +1782,55 @@ final class _EvaluateVisitor
     }
   }
 
-  Value? visitIncludeRule(IncludeRule node) {
-    var nodeWithSpan = AstNode.fake(() => node.spanWithoutContent);
-    var mixin = _addExceptionSpan(node,
-        () => _environment.getMixin(node.name, namespace: node.namespace));
+  /// Evaluate a given [mixin] with [arguments] and [contentCallable]
+  void _applyMixin(
+      Callable? mixin,
+      UserDefinedCallable<Environment>? contentCallable,
+      ArgumentInvocation arguments,
+      AstNode nodeWithSpan,
+      AstNode nodeWithSpanWithoutContent) {
     switch (mixin) {
       case null:
-        throw _exception("Undefined mixin.", node.span);
+        throw _exception("Undefined mixin.", nodeWithSpan.span);
 
-      case BuiltInCallable() when node.content != null:
-        throw _exception("Mixin doesn't accept a content block.", node.span);
-
+      case BuiltInCallable(acceptsContent: false) when contentCallable != null:
+        {
+          var evaluated = _evaluateArguments(arguments);
+          var (overload, _) = mixin.callbackFor(
+              evaluated.positional.length, MapKeySet(evaluated.named));
+          throw MultiSpanSassRuntimeException(
+              "Mixin doesn't accept a content block.",
+              nodeWithSpanWithoutContent.span,
+              "invocation",
+              {overload.spanWithName: "declaration"},
+              _stackTrace(nodeWithSpanWithoutContent.span));
+        }
       case BuiltInCallable():
-        _runBuiltInCallable(node.arguments, mixin, nodeWithSpan);
+        _environment.withContent(contentCallable, () {
+          _environment.asMixin(() {
+            _runBuiltInCallable(arguments, mixin, nodeWithSpanWithoutContent);
+          });
+        });
 
       case UserDefinedCallable<Environment>(
             declaration: MixinRule(hasContent: false)
           )
-          when node.content != null:
+          when contentCallable != null:
         throw MultiSpanSassRuntimeException(
             "Mixin doesn't accept a content block.",
-            node.spanWithoutContent,
+            nodeWithSpanWithoutContent.span,
             "invocation",
             {mixin.declaration.arguments.spanWithName: "declaration"},
-            _stackTrace(node.spanWithoutContent));
+            _stackTrace(nodeWithSpanWithoutContent.span));
 
       case UserDefinedCallable<Environment>():
-        var contentCallable = node.content.andThen((content) =>
-            UserDefinedCallable(content, _environment.closure(),
-                inDependency: _inDependency));
-        _runUserDefinedCallable(node.arguments, mixin, nodeWithSpan, () {
+        _runUserDefinedCallable(arguments, mixin, nodeWithSpanWithoutContent,
+            () {
           _environment.withContent(contentCallable, () {
             _environment.asMixin(() {
               for (var statement in mixin.declaration.children) {
-                _addErrorSpan(nodeWithSpan, () => statement.accept(this));
+                _addErrorSpan(
+                    nodeWithSpanWithoutContent, () => statement.accept(this));
               }
             });
           });
@@ -1772,6 +1839,20 @@ final class _EvaluateVisitor
       case _:
         throw UnsupportedError("Unknown callable type $mixin.");
     }
+  }
+
+  Value? visitIncludeRule(IncludeRule node) {
+    var mixin = _addExceptionSpan(node,
+        () => _environment.getMixin(node.name, namespace: node.namespace));
+    var contentCallable = node.content.andThen((content) => UserDefinedCallable(
+        content, _environment.closure(),
+        inDependency: _inDependency));
+
+    var nodeWithSpanWithoutContent =
+        AstNode.fake(() => node.spanWithoutContent);
+
+    _applyMixin(mixin, contentCallable, node.arguments, node,
+        nodeWithSpanWithoutContent);
 
     return null;
   }